chore: move schema to utils
